walle-prototype/requirements.txt
openai>=1.0.0
SpeechRecognition>=3.10.0
PyAudio>=0.2.13
python-dotenv>=1.0.0
mcp>=1.0.0
requests>=2.32.0
